rename manual_tests as spec
this folder contains the automated specification of tacos-gha
