+++ conftest.py
@@ -13,7 +13,7 @@ from lib.types import Environ
 from lib.types import Generator
 from lib.types import OSPath
 
-TEST_HOME = TACOS_GHA_HOME / "manual_tests"
+TEST_HOME = TACOS_GHA_HOME / "spec"
 
 
 @fixture
