+++ spec/behaviors/force_unlock_on_happy_path.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 
 from lib.parse import Parse
-from manual_tests.lib import tacos_demo
+from spec.lib import tacos_demo
 
 
 def test(pr: tacos_demo.PR) -> None:
