+++ spec/behaviors/message_on_lock_failure.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 
 import pytest
 
-from manual_tests.lib import tacos_demo
-from manual_tests.lib.gh import gh
-from manual_tests.lib.slice import Slices
-from manual_tests.lib.xfail import XFailed
+from spec.lib import tacos_demo
+from spec.lib.gh import gh
+from spec.lib.slice import Slices
+from spec.lib.xfail import XFailed
 
 MESSAGE = "lock failed, on slice prod/slice-3-vm, due to user1, PR #334 "
