+++ spec/lib/slice_test.py
