+++ spec/scenarios/draft_pr.py
@@ -2,9 +2,9 @@
 from __future__ import annotations
 
 from lib.sh import sh
-from manual_tests.lib import tacos_demo
-from manual_tests.lib.gh import gh
-from manual_tests.lib.slice import Slices
+from spec.lib import tacos_demo
+from spec.lib.gh import gh
+from spec.lib.slice import Slices
 
 
 def test(
